Prevent PROJ4 from being rebuilt every time
This was caused by the fact that the GDAL configure script builds a small PROJ4-
based program in order to test that PROJ4 is available. That meant that PROJ4
needed to be built once natively in order to make the GDAL configure script
happy, and once via Emscripten as part of the main build process.